AWS_ETL_Pipeline/requirements.txt
pandas
boto3
psycopg2-binary
